Split Project Euler solutions into separate classes. Extracted common functionality like prime numbers and fibonacci numbers.
