+++ src/main/java/javaslang/collection/Seq.java
@@ -5,6 +5,7 @@
  */
 package javaslang.collection;
 
+import javaslang.Function1;
 import javaslang.Tuple2;
 import javaslang.control.Option;
