+++ src/test/java/javaslang/collection/euler/Euler02.java
@@ -1,0 +1,37 @@
+/*     / \____  _    ______   _____ / \____   ____  _____
+ *    /  \__  \/ \  / \__  \ /  __//  \__  \ /    \/ __  \   Javaslang
+ *  _/  // _\  \  \/  / _\  \\_  \/  // _\  \  /\  \__/  /   Copyright 2014-2015 Daniel Dietrich
+ * /___/ \_____/\____/\_____/____/\___\_____/_/  \_/____/    Licensed under the Apache License, Version 2.0
+ */
+package javaslang.collection.euler;
+
+import javaslang.collection.Stream;
+import org.junit.Test;
+
+import static org.assertj.core.api.StrictAssertions.assertThat;
+
+public class Euler02 {
+
+    /**
+     * <strong>Problem 2: Even Fibonacci numbers</strong>
+     * <p>Each new term in the Fibonacci sequence is generated by adding the previous
+     * two terms. By starting with 1 and 2, the first 10 terms will be: 1, 2, 3, 5,
+     * 8, 13, 21, 34, 55, 89, ...</p>
+     * <p>By considering the terms in the Fibonacci sequence whose values do not exceed
+     * four million, find the sum of the even-valued terms.</p>
+     * <p>See also <a href="https://projecteuler.net/problem=2">projecteuler.net problem 2</a>.</p>
+     */
+    @Test
+    public void shouldSolveProblem2() {
+        assertThat(sumOfEvenFibonacciValuesNotExceeding(90)).isEqualTo(2 + 8 + 34);
+        assertThat(sumOfEvenFibonacciValuesNotExceeding(4_000_000)).isEqualTo(4_613_732);
+    }
+
+    private static long sumOfEvenFibonacciValuesNotExceeding(final int max) {
+        return Stream.from(2)
+                .map(FibonacciNumbers::fibonacci)
+                .takeWhile(f -> f <= max)
+                .filter(f -> f % 2 == 0)
+                .sum().longValue();
+    }
+}
