+++ src/test/java/javaslang/collection/euler/Euler04.java
@@ -1,0 +1,45 @@
+/*     / \____  _    ______   _____ / \____   ____  _____
+ *    /  \__  \/ \  / \__  \ /  __//  \__  \ /    \/ __  \   Javaslang
+ *  _/  // _\  \  \/  / _\  \\_  \/  // _\  \  /\  \__/  /   Copyright 2014-2015 Daniel Dietrich
+ * /___/ \_____/\____/\_____/____/\___\_____/_/  \_/____/    Licensed under the Apache License, Version 2.0
+ */
+package javaslang.collection.euler;
+
+import javaslang.collection.Stream;
+import org.junit.Test;
+
+import static org.assertj.core.api.StrictAssertions.assertThat;
+
+public class Euler04 {
+
+    /**
+     * <strong>Problem 4: Largest palindrome product</strong>
+     * A palindromic number reads the same both ways. The largest palindrome made
+     * from the product of two 2-digit numbers is 9009 = 91 × 99.
+     * <p>
+     * Find the largest palindrome made from the product of two 3-digit numbers.
+     * <p>
+     * See also <a href="https://projecteuler.net/problem=4">projecteuler.net problem 4</a>.
+     */
+    @Test
+    public void shouldSolveProblem4() {
+        assertThat(largestPalindromeOfProductsFromFactorsInRange(10, 99)).isEqualTo(9009);
+        assertThat(largestPalindromeOfProductsFromFactorsInRange(100, 999)).isEqualTo(906609);
+    }
+
+    private static int largestPalindromeOfProductsFromFactorsInRange(final int min, final int max) {
+        return Stream.rangeClosed(min, max)
+                .flatMap(i -> Stream.rangeClosed(i, max).map(j -> i * j))
+                .filter(Euler04::isPalindrome)
+                .max().get();
+    }
+
+    private static boolean isPalindrome(int val) {
+        final String valAsString = Long.toString(val);
+        return valAsString.equals(reverse(valAsString));
+    }
+
+    private static String reverse(String s) {
+        return new StringBuilder(s).reverse().toString();
+    }
+}
